// website/src/types/events.tsx
export type PastEvent = {
  id: number;
  name: string;
  date: string;
  startTime: string;
  endTime: string;
  location: string;
  description: string;
}

export type UpcomingEvent ={
  id: number;
  name: string;
  date: string;
  startTime: string;
  endTime: string;
  location: string;
  description: string;
}